requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
python-dateutil==2.8.2
pandas==2.2.0
scipy==1.12.0
python-dotenv==1.0.1
httpx==0.26.0
markdown==3.5.2
duckdb==0.10.0
Pillow==10.2.0
beautifulsoup4==4.12.3
python-multipart==0.0.9
faker